+++ CHANGELOG.md
@@ -4,6 +4,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0]
+
+### Added
+
+* `preauthscan` command to scan for accounts that do not require Kerberos pre-authentication
+* `/preauth` argument to the `kerberoast` command, to kerberoast with an account that does not require Kerberos pre-authentication
+* `/nopreauth` flag to the `asktgt` command, to request a TGT without providing pre-authentication
+* `/service` argument to the `asktgt` command, to request service tickets using an AS-REQ
+
+## [2.1.0]
+
+### Added
+
+* `logonsession` command to list information about the current logon session
+
+## [2.0.3]
+
+### Added
+
+* A `/debug` flag that outputs base64 encodings of the inputs to/outputs from the ASN.1 decoding/encoding functions
+* `/createnetonly` parameter to S4U (@tyranid)
+* `/ticket` option to createnetonly to import a ticket into the new process without requiring privileges (@tyranid)
+
+### Fixed
+
+* Handling for KERB_ERRORs
+  
 ## [2.0.2]
 
 ### Added